refactor: change plugin name to flutterForge
